requirements.txt
gradio>=4.0.0
openai>=1.0.0
google-generativeai>=0.3.0
transformers>=4.30.0
torch>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
markdown>=3.4.0
weasyprint>=60.0
reportlab>=4.0.0
pyperclip>=1.8.0

